+++ src/editor/core/plugins/image/plugin.js
@@ -14,7 +14,7 @@ const imagePlugin = (context) => {
     context.editor.$.on('mouseleave', e => {
         let target =  e.toElement || e.relatedTarget;
         if(!$(target).closest('.humhub-richtext-inline-menu').length) {
-            $('.humhub-richtext-inline-menu').remove();
+          $('.humhub-richtext-inline-menu').remove();
         }
     });
 
@@ -26,7 +26,7 @@ const imagePlugin = (context) => {
         },
         filterTransaction: (tr, state) => {
             if(!(tr.curSelection instanceof NodeSelection)) {
-                $('.humhub-richtext-image-edit').remove();
+             $('.humhub-richtext-image-edit').remove();
             }
 
             return true;
@@ -39,18 +39,28 @@ class ImageView {
         // The editor will use this as the node's DOM representation
         this.createDom(node);
 
+        context.event.on('clear, serialize', function() {
+           $('.humhub-richtext-inline-menu').remove();
+        });
+
 
         this.dom.addEventListener("mouseenter", e => {
             let $img = $(this.dom);
             let offset = $img.offset();
+            let editorOffset = context.editor.$.offset();
+
+            if(offset.top < editorOffset.top) {
+                return;
+            }
 
             let $edit = $('<div>').addClass('humhub-richtext-inline-menu').addClass('humhub-richtext-image-edit')
                 .html('<button class="btn btn-primary btn-xs"><i class="fa fa-pencil"></i></button>')
                 .css({
                     position: 'absolute',
                     left: offset.left + $img.width() - (25),
-                    top: offset.top + 5
-                }).on('click', (evt) => {
+                    top: offset.top + 5,
+                    'z-index': 997
+                }).on('mousedown', (evt) => {
                     let view = context.editor.view;
                     let doc = view.state.doc;
                     view.dispatch(view.state.tr.setSelection(NodeSelection.create(doc, view.posAtDOM(this.dom))).scrollIntoView());